Move common properties providers in the common module
