+++ core/common/src/main/kotlin/com/merxury/blocker/core/di/AppPackageName.kt
@@ -14,14 +14,10 @@
  * limitations under the License.
  */
 
-package com.merxury.blocker.core.data.di
+package com.merxury.blocker.core.di
 
 import javax.inject.Qualifier
 
 @Qualifier
 @Retention(AnnotationRetention.BINARY)
-annotation class FilesDir
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class CacheDir
+annotation class AppPackageName
