+++ core/data/src/main/kotlin/com/merxury/blocker/core/data/respository/componentdetail/datasource/LocalComponentDetailDataSource.kt
@@ -16,9 +16,9 @@
 
 package com.merxury.blocker.core.data.respository.componentdetail.datasource
 
-import com.merxury.blocker.core.data.di.FilesDir
 import com.merxury.blocker.core.data.di.RuleBaseFolder
 import com.merxury.blocker.core.data.respository.userdata.UserDataRepository
+import com.merxury.blocker.core.di.FilesDir
 import com.merxury.blocker.core.dispatchers.BlockerDispatchers.IO
 import com.merxury.blocker.core.dispatchers.Dispatcher
 import com.merxury.blocker.core.model.data.ComponentDetail
