+++ core/data/src/main/kotlin/com/merxury/blocker/core/data/respository/componentdetail/datasource/UserGeneratedComponentDetailDataSource.kt
@@ -16,10 +16,10 @@
 
 package com.merxury.blocker.core.data.respository.componentdetail.datasource
 
-import com.merxury.blocker.core.data.di.FilesDir
 import com.merxury.blocker.core.data.di.GeneratedRuleBaseFolder
 import com.merxury.blocker.core.data.respository.component.CacheComponentDataSource
 import com.merxury.blocker.core.data.respository.userdata.UserDataRepository
+import com.merxury.blocker.core.di.FilesDir
 import com.merxury.blocker.core.dispatchers.BlockerDispatchers.IO
 import com.merxury.blocker.core.dispatchers.Dispatcher
 import com.merxury.blocker.core.model.data.ComponentDetail
