+++ core/data/src/main/kotlin/com/merxury/blocker/core/data/respository/generalrule/LocalGeneralRuleDataSource.kt
@@ -16,8 +16,8 @@
 
 package com.merxury.blocker.core.data.respository.generalrule
 
-import com.merxury.blocker.core.data.di.FilesDir
 import com.merxury.blocker.core.data.di.RuleBaseFolder
+import com.merxury.blocker.core.di.FilesDir
 import com.merxury.blocker.core.dispatchers.BlockerDispatchers.IO
 import com.merxury.blocker.core.dispatchers.Dispatcher
 import com.merxury.blocker.core.model.data.GeneralRule
