+++ core/domain/src/main/kotlin/com/merxury/blocker/core/domain/ZipAllRuleUseCase.kt
@@ -16,10 +16,10 @@
 
 package com.merxury.blocker.core.domain
 
-import com.merxury.blocker.core.data.di.CacheDir
-import com.merxury.blocker.core.data.di.FilesDir
 import com.merxury.blocker.core.data.di.GeneratedRuleBaseFolder
 import com.merxury.blocker.core.data.respository.userdata.UserDataRepository
+import com.merxury.blocker.core.di.CacheDir
+import com.merxury.blocker.core.di.FilesDir
 import com.merxury.blocker.core.dispatchers.BlockerDispatchers.IO
 import com.merxury.blocker.core.dispatchers.Dispatcher
 import com.merxury.blocker.core.domain.model.ZippedRule
